+++ Niconicome/Models/Domain/Niconico/Watch/WatchInfohandler.cs
@@ -8,6 +8,7 @@ using AngleSharp.Html.Dom;
 using Niconicome.Extensions.System;
 using Niconicome.Models.Domain.Niconico.Net.Html;
 using Niconicome.Models.Domain.Niconico.Net.Json;
+using Niconicome.Models.Domain.Niconico.Video.Infomations;
 using Niconicome.Models.Domain.Utils;
 using DmcRequest = Niconicome.Models.Domain.Niconico.Net.Json.WatchPage.DMC.Request;
 using WatchJson = Niconicome.Models.Domain.Niconico.Net.Json.WatchPage.V2;
@@ -18,72 +19,6 @@ namespace Niconicome.Models.Domain.Niconico.Watch
     {
         Task<IDomainVideoInfo> GetVideoInfoAsync(string id, WatchInfoOptions options);
         WatchInfoHandlerState State { get; }
-    }
-
-    public interface IDomainVideoInfo
-    {
-        string Title { get; set; }
-        string Id { get; set; }
-        string ChannelID { get; set; }
-        string ChannelName { get; set; }
-        string Owner { get; set; }
-        int OwnerID { get; set; }
-        int ViewCount { get; }
-        int MylistCount { get; }
-        int CommentCount { get; }
-        int LikeCount { get; set; }
-        int Duration { get; set; }
-        IEnumerable<string> Tags { get; set; }
-        IDmcInfo DmcInfo { get; set; }
-    }
-
-    public interface IDmcInfo
-    {
-        string Title { get; set; }
-        string Id { get; set; }
-        string Owner { get; set; }
-        int OwnerID { get; set; }
-        string Userkey { get; }
-        string UserId { get; }
-        string ChannelID { get; }
-        string ChannelName { get; }
-        string Description { get; }
-        int ViewCount { get; }
-        int CommentCount { get; }
-        int MylistCount { get; }
-        int LikeCount { get; }
-        int Duration { get; }
-        IEnumerable<string> Tags { get; set; }
-        bool IsDownloadsble { get; set; }
-        bool IsEncrypted { get; set; }
-        bool IsOfficial { get; set; }
-        DateTime UploadedOn { get; set; }
-        DateTime DownloadStartedOn { get; set; }
-        IThumbInfo ThumbInfo { get; }
-        ISessionInfo SessionInfo { get; }
-        List<WatchJson::Thread> CommentThreads { get; }
-    }
-
-    public interface IThumbInfo
-    {
-        string? Large { get; set; }
-        string? Normal { get; set; }
-    }
-
-    public interface ISessionInfo
-    {
-        string? RecipeId { get; set; }
-        string? ContentId { get; set; }
-        DmcRequest::Content_Src_Id_Sets ContentSrcIdSets { get; set; }
-        int HeartbeatLifetime { get; set; }
-        string? Token { get; set; }
-        string? Signature { get; set; }
-        string? AuthType { get; set; }
-        int ContentKeyTimeout { get; set; }
-        string? ServiceUserId { get; set; }
-        string? PlayerId { get; set; }
-        string? TransferPriset { get; set; }
-        double Priority { get; set; }
     }
     public interface IWatchPageHtmlParser
     {
@@ -379,211 +314,6 @@ namespace Niconicome.Models.Domain.Niconico.Watch
             return sets;
         }
     }
-
-    /// <summary>
-    /// 動画情報(ルート)
-    /// </summary>
-    public class DomainVideoInfo : IDomainVideoInfo
-    {
-        /// <summary>
-        /// 動画タイトル
-        /// </summary>
-        public string Title { get; set; } = string.Empty;
-
-        /// <summary>
-        /// 動画ID
-        /// </summary>
-        public string Id { get; set; } = string.Empty;
-
-        /// <summary>
-        /// ファイル名
-        /// </summary>
-        public string FileName { get; set; } = string.Empty;
-
-        /// <summary>
-        /// チャンネルID
-        /// </summary>
-        public string ChannelID { get; set; } = string.Empty;
-
-        /// <summary>
-        /// チャンネル名
-        /// </summary>
-        public string ChannelName { get; set; } = string.Empty;
-
-        /// <summary>
-        /// 投稿者名
-        /// </summary>
-        public string Owner { get; set; } = string.Empty;
-
-        /// <summary>
-        /// 投稿者ユーザーID
-        /// </summary>
-        public int OwnerID { get; set; }
-
-        /// <summary>
-        /// 再生回数
-        /// </summary>
-        public int ViewCount { get; set; }
-
-        /// <summary>
-        /// マイリス数
-        /// </summary>
-        public int MylistCount { get; set; }
-
-        /// <summary>
-        /// コメント数
-        /// </summary>
-        public int CommentCount { get; set; }
-
-        /// <summary>
-        /// いいね数
-        /// </summary>
-        public int LikeCount { get; set; }
-
-        /// <summary>
-        /// 再生時間
-        /// </summary>
-        public int Duration { get; set; }
-
-
-        /// <summary>
-        /// タグ
-        /// </summary>
-        public IEnumerable<string> Tags { get; set; } = new List<string>();
-
-        /// <summary>
-        /// DMC情報
-        /// </summary>
-        public IDmcInfo DmcInfo { get; set; } = new DmcInfo();
-
-
-    }
-
-    /// <summary>
-    /// APi等へのアクセスに必要な情報を格納する
-    /// </summary>
-    public class DmcInfo : IDmcInfo
-    {
-        public string Title { get; set; } = string.Empty;
-
-        public string Id { get; set; } = string.Empty;
-
-        public string Owner { get; set; } = string.Empty;
-
-        public string Userkey { get; set; } = string.Empty;
-
-        public string UserId { get; set; } = string.Empty;
-
-        public string ChannelID { get; set; } = string.Empty;
-
-        public string ChannelName { get; set; } = string.Empty;
-
-        public string Description { get; set; } = string.Empty;
-
-
-        public int ViewCount { get; set; }
-
-        public int CommentCount { get; set; }
-
-        public int LikeCount { get; set; }
-
-        public int MylistCount { get; set; }
-
-        public int Duration { get; set; }
-
-        public int OwnerID { get; set; }
-
-        public IEnumerable<string> Tags { get; set; } = new List<string>();
-
-        /// <summary>
-        /// ダウンロード可能フラグ
-        /// </summary>
-        public bool IsDownloadsble { get; set; } = true;
-
-        /// <summary>
-        /// 暗号化フラグ
-        /// </summary>
-       　public bool IsEncrypted { get; set; }
-
-        /// <summary>
-        /// 公式動画フラグ
-        /// </summary>
-        public bool IsOfficial { get; set; }
-
-
-        /// <summary>
-        /// 投稿日時
-        /// </summary>
-        public DateTime UploadedOn { get; set; }
-
-        /// <summary>
-        /// DL開始日時
-        /// </summary>
-        public DateTime DownloadStartedOn { get; set; }
-
-
-        /// <summary>
-        /// サムネイル
-        /// </summary>
-        public IThumbInfo ThumbInfo { get; private set; } = new ThumbInfo();
-
-        /// <summary>
-        /// セッション情報
-        /// </summary>
-        public ISessionInfo SessionInfo { get; private set; } = new SessionInfo();
-
-        /// <summary>
-        /// コメントスレッド
-        /// </summary>
-        public List<WatchJson::Thread> CommentThreads { get; set; } = new();
-    }
-
-    /// <summary>
-    /// セッション情報
-    /// </summary>
-    public class SessionInfo : ISessionInfo
-    {
-        public string? RecipeId { get; set; }
-
-        public string? ContentId { get; set; }
-
-        public DmcRequest::Content_Src_Id_Sets ContentSrcIdSets { get; set; } = new();
-
-        public int HeartbeatLifetime { get; set; }
-
-        public string? Token { get; set; }
-
-        public string? Signature { get; set; }
-
-        public string? AuthType { get; set; }
-
-        public string? TransferPriset { get; set; }
-
-        public int ContentKeyTimeout { get; set; }
-
-        public string? ServiceUserId { get; set; }
-
-        public string? PlayerId { get; set; }
-
-        public double Priority { get; set; }
-    }
-
-    /// <summary>
-    /// サムネイル情報
-    /// </summary>
-    public class ThumbInfo : IThumbInfo
-    {
-        /// <summary>
-        /// 大サムネイル
-        /// </summary>
-        public string? Large { get; set; }
-
-        /// <summary>
-        /// 通常
-        /// </summary>
-        public string? Normal { get; set; }
-    }
-
 
 
     /// <summary>
