+++ Niconicome/Models/Domain/Utils/PathOrganizer.cs
@@ -4,13 +4,13 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Watch = Niconicome.Models.Domain.Niconico.Watch;
+using Niconicome.Models.Domain.Niconico.Video.Infomations;
 
 namespace Niconicome.Models.Domain.Utils
 {
     public interface IPathOrganizer
     {
-        string GetFIlePath(string format, Watch.IDmcInfo dmcInfo, string extension, string folderName, bool replaceStricted, bool overWrite, string? suffix = null);
+        string GetFIlePath(string format, IDmcInfo dmcInfo, string extension, string folderName, bool replaceStricted, bool overWrite, string? suffix = null);
     }
 
     public class PathOrganizer : IPathOrganizer
@@ -24,7 +24,7 @@ namespace Niconicome.Models.Domain.Utils
         private readonly INiconicoUtils niconicoUtils;
         #endregion
 
-        public string GetFIlePath(string format, Watch::IDmcInfo dmcInfo, string extension, string folderName, bool replaceStricted, bool overWrite, string? suffix = null)
+        public string GetFIlePath(string format, IDmcInfo dmcInfo, string extension, string folderName, bool replaceStricted, bool overWrite, string? suffix = null)
         {
             var filename = this.niconicoUtils.GetFileName(format, dmcInfo, extension, replaceStricted, suffix);
             var filePath = Path.Combine(folderName, filename);
