+++ Niconicome/Models/Network/Watch/Watch.cs
@@ -2,32 +2,13 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Niconicome.Extensions.System;
+using Niconicome.Models.Domain.Niconico.Video.Infomations;
 using Niconicome.Models.Domain.Niconico.Watch;
 using Niconicome.Models.Domain.Utils;
 using Niconicome.Models.Playlist;
 
 namespace Niconicome.Models.Network.Watch
 {
-    public interface IVideoInfo
-    {
-        string Title { get; set; }
-        string Id { get; set; }
-        string FileName { get; set; }
-        string ChannelID { get; set; }
-        string ChannelName { get; set; }
-        string OwnerName { get; set; }
-        int ViewCount { get; set; }
-        int CommentCount { get; set; }
-        int MylistCount { get; set; }
-        int LikeCount { get; set; }
-        int OwnerID { get; set; }
-        int Duration { get; set; }
-        IEnumerable<string> Tags { get; set; }
-        DateTime UploadedOn { get; set; }
-        Uri LargeThumbUri { get; set; }
-        Uri ThumbUri { get; set; }
-        IListVideoInfo ConvertToTreeVideoInfo();
-    }
 
     public interface IWatch
     {
@@ -122,73 +103,6 @@ namespace Niconicome.Models.Network.Watch
 
 
             return result;
-        }
-
-    }
-
-    /// <summary>
-    /// 動画情報
-    /// </summary>
-    public class VIdeoInfo : IVideoInfo
-    {
-        public static string DeletedVideoThumb { get; private set; } = "https://nicovideo.cdn.nimg.jp/web/img/common/video_deleted.jpg";
-
-        public string Title { get; set; } = string.Empty;
-
-        public string Id { get; set; } = string.Empty;
-
-        public string FileName { get; set; } = string.Empty;
-
-        public string ChannelID { get; set; } = string.Empty;
-
-        public string ChannelName { get; set; } = string.Empty;
-
-        public string OwnerName { get; set; } = string.Empty;
-
-        public int ViewCount { get; set; }
-
-        public int CommentCount { get; set; }
-
-        public int MylistCount { get; set; }
-
-        public int LikeCount { get; set; }
-
-        public int OwnerID { get; set; }
-
-        public int Duration { get; set; }
-
-        public IEnumerable<string> Tags { get; set; } = new List<string>();
-
-        public DateTime UploadedOn { get; set; }
-
-        public Uri LargeThumbUri { get; set; } = new Uri(DeletedVideoThumb);
-
-        public Uri ThumbUri { get; set; } = new Uri(DeletedVideoThumb);
-
-        /// <summary>
-        /// Viewから参照可能な形式に変換する
-        /// </summary>
-        /// <returns></returns>
-        public IListVideoInfo ConvertToTreeVideoInfo()
-        {
-            var video = new NonBindableListVideoInfo();
-            video.Title.Value = this.Title;
-            video.NiconicoId.Value = this.Id;
-            video.UploadedOn.Value = this.UploadedOn;
-            video.LargeThumbUrl.Value = this.LargeThumbUri.AbsoluteUri;
-            video.ThumbUrl.Value = this.ThumbUri.AbsoluteUri;
-            video.Tags = this.Tags;
-            video.ViewCount.Value = this.ViewCount;
-            video.FileName.Value = this.FileName;
-            video.ChannelID.Value = this.ChannelID;
-            video.ChannelName.Value = this.ChannelName;
-            video.MylistCount.Value = this.MylistCount;
-            video.CommentCount.Value = this.CommentCount;
-            video.LikeCount.Value = this.LikeCount;
-            video.OwnerID.Value = this.OwnerID;
-            video.Duration.Value = this.Duration;
-            video.OwnerName.Value = this.OwnerName;
-            return video;
         }
 
     }
